requirements.txt
alembic==0.9.9
asn1crypto==0.24.0
Babel==2.6.0
bcrypt==3.1.4
blinker==1.4
cffi==1.11.5
click==6.7
cryptography==2.2.2
dominate==2.3.1
Flask==1.0.2
Flask-Admin==1.5.1
Flask-BabelEx==0.9.3
Flask-Bootstrap==3.3.7.1
Flask-Images==2.1.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-MySQLdb==0.2.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-User==1.0.1.4
Flask-WTF==0.14.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
mysqlclient==1.3.12
passlib==1.7.1
PIL==1.1.7
Pillow==5.1.0
pillowcase==2.0.0
pycparser==2.18
PyJWT==1.6.4
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.4
six==1.11.0
speaklater==1.3
SQLAlchemy==1.2.8
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
